+++ apps/forge-cli/src/cli/index.js
@@ -1,8 +1,9 @@
 exports.account = require('./account');
-exports.project = require('./project');
 exports.asset = require('./asset');
 exports.block = require('./block');
 exports.misc = require('./misc');
 exports.node = require('./node');
+exports.project = require('./project');
+exports.protocol = require('./protocol');
 exports.release = require('./release');
 exports.tx = require('./tx');