+++ packages/forge-message/index.d.ts
@@ -1,4 +1,4 @@
-// Generate by [js2dts@0.3.2](https://github.com/whxaxes/js2dts#readme)
+// Generate by [js2dts@0.3.3](https://github.com/whxaxes/js2dts#readme)
 
 /**
  * Format an message from RPC to UI friendly
@@ -130,6 +130,15 @@ declare function attachFormatFn(type: string, data: any, key?: string): void;
  * @param {string} type
  */
 declare function attachExampleFn(type: string, host: any, key: any): void;
+/**
+ * Add type provider that can be used to format/create messages
+ *
+ * @param {object} provider - proto generated from {@see @arcblock/forge-proto}
+ */
+declare function addProvider(provider: any): void;
+declare function getMessageType(type: any): any;
+declare function toTypeUrl(type: any): any;
+declare function fromTypeUrl(url: any): any;
 declare const _ArcblockForgeMessage: _ArcblockForgeMessage.T101;
 declare namespace _ArcblockForgeMessage {
   export interface T100 {
@@ -148,6 +157,10 @@ declare namespace _ArcblockForgeMessage {
     decodeBigInt: typeof decodeBigInt;
     attachFormatFn: typeof attachFormatFn;
     attachExampleFn: typeof attachExampleFn;
+    addProvider: typeof addProvider;
+    getMessageType: typeof getMessageType;
+    toTypeUrl: typeof toTypeUrl;
+    fromTypeUrl: typeof fromTypeUrl;
   }
 }
 export = _ArcblockForgeMessage;