+++ packages/forge-proto/provider.js
@@ -9,10 +9,19 @@
  */
 
 const get = require('lodash.get');
-const json = require('./lib/spec.json');
+const debug = require('debug')(require('./package.json').name);
 
-module.exports = function(proto) {
-  const { types, services, vendorTypes } = proto;
+/**
+ * Generate type provider from types map and json spec
+ *
+ * @param {object} proto - collection of types
+ * @param {object} json - collection of fields/types used in the types map
+ * @param {object} urls - collection of typeUrls registered to forge-core
+ * @returns {object}
+ */
+module.exports = function(proto, json, urls = {}) {
+  debug('provider.create', { urls });
+  const { types, services = {}, vendorTypes = {} } = proto;
 
   const txTypePattern = /Tx$/;
   const stateTypePattern = /State$/;
@@ -80,7 +89,7 @@ module.exports = function(proto) {
     }, {});
 
   // extract typeUrls
-  const typeUrls = createTypeUrls(abi);
+  const typeUrls = Object.assign(createTypeUrls(abi), urls || {});
 
   // extract transactions and stakes
   enums.SupportedTxs = Object.keys(types).filter(
@@ -103,6 +112,9 @@ module.exports = function(proto) {
         }
         if (['TransactionInfo', 'TxStatus'].includes(type)) {
           typeUrl = `fg:x:${lowerUnder(type)}`;
+        }
+        if (type === 'AssetFactory') {
+          typeUrl = 'fg:x:asset_factory';
         }
         if (type === 'DummyCodec') {
           typeUrl = 'fg:x:address';